+++ lib/enum.dart
@@ -1,0 +1,3 @@
+
+
+enum Shape { CIRCLE, SQUARE }