+++ lib/power_switch_button.dart
@@ -3,19 +3,20 @@ library power_switch_button;
 import 'package:flutter/material.dart';
 
 import 'dashed_circle_painter.dart';
+import 'enum.dart';
 
 /// A customizable switch button with animated on/off states and dashed border.
 class PowerSwitchButton extends StatefulWidget {
   /// The size of the switch button.
   final double size;
 
-  /// The stroke width of the dashed circle.
+  /// The stroke width of the dashed border.
   final double strokeWidth;
 
-  /// The width of each dash in the dashed circle.
+  /// The width of each dash in the dashed border.
   final double dashWidth;
 
-  /// The space between each dash in the dashed circle.
+  /// The space between each dash in the dashed border.
   final double dashSpace;
 
   /// The color of the switch when it is in the 'on' position.
@@ -29,6 +30,33 @@ class PowerSwitchButton extends StatefulWidget {
 
   /// The color of the power icon.
   final Color iconColor;
+
+  /// The custom icon for the switch.
+  final IconData? customIcon;
+
+  /// The label text to display inside the switch.
+  final String? label;
+
+  /// Whether the switch is disabled.
+  final bool disabled;
+
+  /// The gradient colors for the 'on' state.
+  final Gradient? onGradient;
+
+  /// The gradient colors for the 'off' state.
+  final Gradient? offGradient;
+
+  /// The duration of the toggle animation.
+  final Duration animationDuration;
+
+  /// The curve of the toggle animation.
+  final Curve animationCurve;
+
+  /// The shape of the switch button.
+  final Shape shape;
+
+  /// The border radius for the square shape. Optional.
+  final double? borderRadius;
 
   /// Callback that gets called when the switch is toggled.
   final ValueChanged<bool> onToggle;
@@ -44,6 +72,15 @@ class PowerSwitchButton extends StatefulWidget {
     required this.backgroundColor,
     required this.iconColor,
     required this.onToggle,
+    this.customIcon,
+    this.label,
+    this.disabled = false,
+    this.onGradient,
+    this.offGradient,
+    this.shape = Shape.CIRCLE, // Default shape
+    this.borderRadius, // Optional border radius
+    this.animationDuration = const Duration(milliseconds: 300),
+    this.animationCurve = Curves.easeInOut,
   });
 
   @override
@@ -64,16 +101,18 @@ class _PowerSwitchButtonState extends State<PowerSwitchButton>
       vsync: this,
     );
     _scaleAnimation = Tween<double>(begin: 1.0, end: 0.9).animate(
-      CurvedAnimation(parent: _controller, curve: Curves.easeInOut),
+      CurvedAnimation(parent: _controller, curve: widget.animationCurve),
     );
   }
 
   /// Toggles the switch state and calls the onToggle callback.
   void _toggleSwitch() {
-    setState(() {
-      isOn = !isOn;
-      widget.onToggle(isOn);
-    });
+    if (!widget.disabled) {
+      setState(() {
+        isOn = !isOn;
+        widget.onToggle(isOn);
+      });
+    }
   }
 
   @override
@@ -90,16 +129,23 @@ class _PowerSwitchButtonState extends State<PowerSwitchButton>
           height: outerCircleSize,
           decoration: BoxDecoration(
             color: widget.backgroundColor,
-            shape: BoxShape.circle,
+            shape: widget.shape == Shape.CIRCLE
+                ? BoxShape.circle
+                : BoxShape.rectangle,
+            borderRadius: widget.shape == Shape.SQUARE && widget.borderRadius != null
+                ? BorderRadius.circular(widget.borderRadius!)
+                : null,
           ),
         ),
         CustomPaint(
           size: Size(middleCircleSize, middleCircleSize),
-          painter: DashedCirclePainter(
+          painter: DashedShapePainter(
             strokeWidth: widget.strokeWidth,
             dashWidth1: widget.dashWidth,
             dashSpace1: widget.dashSpace,
             strokeColor: isOn ? widget.onColor : widget.offColor,
+            shape: widget.shape,
+            borderRadius: widget.borderRadius, // Pass border radius
           ),
         ),
         GestureDetector(
@@ -111,18 +157,48 @@ class _PowerSwitchButtonState extends State<PowerSwitchButton>
           child: ScaleTransition(
             scale: _scaleAnimation,
             child: AnimatedContainer(
-              duration: Duration(milliseconds: 300),
+              duration: widget.animationDuration,
               width: innerCircleSize,
               height: innerCircleSize,
               decoration: BoxDecoration(
-                color: isOn ? widget.onColor : widget.offColor,
-                shape: BoxShape.circle,
+                shape: widget.shape == Shape.CIRCLE
+                    ? BoxShape.circle
+                    : BoxShape.rectangle,
+                borderRadius: widget.shape == Shape.SQUARE && widget.borderRadius != null
+                    ? BorderRadius.circular(widget.borderRadius!)
+                    : null,
+                gradient: isOn ? widget.onGradient : widget.offGradient,
+                color: widget.disabled
+                    ? Colors.grey.withOpacity(0.5)
+                    : isOn
+                    ? widget.onColor
+                    : widget.offColor,
               ),
               child: Center(
-                child: Icon(
-                  Icons.power_settings_new,
-                  size: innerCircleSize * 0.5,
-                  color: widget.iconColor,
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: [
+                    if (widget.customIcon != null)
+                      Icon(
+                        widget.customIcon,
+                        size: innerCircleSize * 0.5,
+                        color: widget.iconColor,
+                      )
+                    else
+                      Icon(
+                        Icons.power_settings_new,
+                        size: innerCircleSize * 0.5,
+                        color: widget.iconColor,
+                      ),
+                    if (widget.label != null)
+                      Text(
+                        widget.label!,
+                        style: TextStyle(
+                          color: widget.iconColor,
+                          fontSize: innerCircleSize * 0.1,
+                        ),
+                      ),
+                  ],
                 ),
               ),
             ),